+++ packages/health/ios/Classes/SwiftHealthPlugin.swift
@@ -465,8 +465,10 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
 
         var nutrition = Set<HKSample>()
 
-        let caloriesSample = HKQuantitySample(type: HKSampleType.quantityType(forIdentifier: .dietaryEnergyConsumed)!, quantity: HKQuantity(unit: HKUnit.kilocalorie(), doubleValue: calories), start: dateFrom, end: dateTo, metadata: metadata)
-        nutrition.insert(caloriesSample)
+        if(calories > 0) {
+            let caloriesSample = HKQuantitySample(type: HKSampleType.quantityType(forIdentifier: .dietaryEnergyConsumed)!, quantity: HKQuantity(unit: HKUnit.kilocalorie(), doubleValue: calories), start: dateFrom, end: dateTo, metadata: metadata)
+            nutrition.insert(caloriesSample)
+        }
 
         if(carbs > 0) {
             let carbsSample = HKQuantitySample(type: HKSampleType.quantityType(forIdentifier: .dietaryCarbohydrates)!, quantity: HKQuantity(unit: HKUnit.gram(), doubleValue: carbs), start: dateFrom, end: dateTo, metadata: metadata)