+++ packages/health/lib/health.dart
@@ -10,7 +10,7 @@ import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
-part 'src/data_types.dart';
+part 'src/heath_data_types.dart';
 part 'src/functions.dart';
 part 'src/health_data_point.dart';
 part 'src/health_value_types.dart';