+++ packages/health/lib/src/workout_summary.dart
@@ -1,9 +1,11 @@
 part of '../health.dart';
 
 /// A [WorkoutSummary] object store vary metrics of a workout.
-///  * totalDistance - The total distance that was traveled during a workout.
-///  * totalEnergyBurned - The amount of energy that was burned during a workout.
-///  * totalSteps - The count of steps was burned during a workout.
+///
+///  * [workoutType] - The type of workout. See [HealthWorkoutActivityType] for available types.
+///  * [totalDistance] - The total distance that was traveled during a workout.
+///  * [totalEnergyBurned] - The amount of energy that was burned during a workout.
+///  * [totalSteps] - The number of steps during a workout.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class WorkoutSummary {
   /// Workout type.
@@ -18,12 +20,21 @@ class WorkoutSummary {
   /// The total steps value of the workout.
   num totalSteps;
 
-  WorkoutSummary(
-    this.workoutType,
-    this.totalDistance,
-    this.totalEnergyBurned,
-    this.totalSteps,
-  );
+  WorkoutSummary({
+    required this.workoutType,
+    required this.totalDistance,
+    required this.totalEnergyBurned,
+    required this.totalSteps,
+  });
+
+  /// Create a [WorkoutSummary] based on a health data point from native data format.
+  factory WorkoutSummary.fromHealthDataPoint(dynamic dataPoint) =>
+      WorkoutSummary(
+        workoutType: dataPoint['workout_type'] as String? ?? '',
+        totalDistance: dataPoint['total_distance'] as num? ?? 0,
+        totalEnergyBurned: dataPoint['total_energy_burned'] as num? ?? 0,
+        totalSteps: dataPoint['total_steps'] as num? ?? 0,
+      );
 
   /// Create a [HealthDataPoint] from json.
   factory WorkoutSummary.fromJson(Map<String, dynamic> json) =>
@@ -31,22 +42,6 @@ class WorkoutSummary {
 
   /// Convert this [HealthDataPoint] to json.
   Map<String, dynamic> toJson() => _$WorkoutSummaryToJson(this);
-
-  // /// Converts a json object to the [WorkoutSummary]
-  // factory WorkoutSummary.fromJson(json) => WorkoutSummary(
-  //       json['workoutType'],
-  //       json['totalDistance'],
-  //       json['totalEnergyBurned'],
-  //       json['totalSteps'],
-  //     );
-
-  // /// Converts the [WorkoutSummary] to a json object
-  // Map<String, dynamic> toJson() => {
-  //       'workoutType': workoutType,
-  //       'totalDistance': totalDistance,
-  //       'totalEnergyBurned': totalEnergyBurned,
-  //       'totalSteps': totalSteps
-  //     };
 
   @override
   String toString() => '$runtimeType - '